Create findDuplicates.cpp
Given an array a[] of size N which contains elements from 0 to N-1, you need to find all the elements occurring more than once in the given array. Return the answer in ascending order.

Note: The extra space is only for the array to be returned.
Try and perform all operations within the provided array. 

https://practice.geeksforgeeks.org/problems/find-duplicates-in-an-array/1